# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateDebtRequest:
    properties:
      account_id:
        type: string
      client_id:
        type: string
      contact_id:
        type: string
      contact_name:
        type: string
      direction:
        description: lent, borrowed
        type: string
      due_date:
        type: string
      interest_rate:
        type: string
      interest_rule:
        type: string
      name:
        type: string
      principal:
        type: string
      start_date:
        type: string
    required:
    - account_id
    - direction
    - due_date
    - principal
    - start_date
    type: object
  dto.CreateGroupExpenseRequest:
    properties:
      account_id:
        type: string
      amount:
        type: string
      category_id:
        type: string
      client_id:
        type: string
      description:
        type: string
      external_ref:
        type: string
      notes:
        type: string
      occurred_at:
        type: string
      occurred_date:
        type: string
      occurred_time:
        type: string
      owner_original_amount:
        type: string
      participants:
        items:
          $ref: '#/definitions/dto.GroupExpenseParticipantRequest'
        type: array
      tag_ids:
        items:
          type: string
        type: array
    type: object
...
  dto.DebtPaymentRequest:
    properties:
      amount_paid:
        description: Total paid, can be split by service
        type: string
      interest_paid:
        type: string
      principal_paid:
        type: string
      transaction_id:
        type: string
    required:
    - transaction_id
    type: object
  dto.GroupExpenseParticipantRequest:
    properties:
      create_debt:
        type: boolean
      name:
        type: string
      original_amount:
        type: string
    type: object
  dto.GroupExpenseSettleRequest:
    properties:
      account_id:
        type: string
      occurred_at:
        type: string
      occurred_date:
        type: string
      occurred_time:
        type: string
    type: object
...
  dto.UpdateDebtRequest:
    properties:
      due_date:
        type: string
      interest_rate:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
...
  entity.Debt:
    properties:
      account_id:
        type: string
      accrued_interest:
        type: string
      client_id:
        type: string
      closed_at:
        type: string
      contact_avatar_url:
        type: string
      contact_id:
        type: string
      contact_name:
        type: string
      created_at:
        type: string
      currency:
        type: string
      direction:
        description: lent, borrowed
        type: string
      due_date:
        type: string
      id:
        type: string
      interest_rate:
        type: string
      interest_rule:
        type: string
      name:
        type: string
      outstanding_principal:
        type: string
      principal:
        type: string
      start_date:
        type: string
      status:
        description: active, paid, cancelled
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  entity.Transaction:
    properties:
      account_currency:
        description: Currencies are derived from linked accounts (not stored on transactions).
        type: string
      account_id:
        type: string
      amount:
        type: string
      category_colors:
        items:
          type: string
        type: array
      category_ids:
        items:
          type: string
        type: array
      category_names:
        description: Wrapper fields (enriched)
        items:
          type: string
        type: array
      client_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      exchange_rate:
        type: string
      external_ref:
        type: string
      from_account_id:
        type: string
      from_amount:
        description: For FX transfers, amounts can differ per side.
        type: string
      from_currency:
        type: string
      id:
        type: string
      line_items:
        items:
          $ref: '#/definitions/entity.TransactionLineItem'
        type: array
      occurred_at:
        type: string
      occurred_date:
        type: string
      status:
        description: pending, posted, cancelled
        type: string
      tag_colors:
        items:
          type: string
        type: array
      tag_ids:
        items:
          type: string
        type: array
      tag_names:
        items:
          type: string
        type: array
      to_account_id:
        type: string
      to_amount:
        type: string
      to_currency:
        type: string
      type:
        description: expense, income, transfer
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  entity.TransactionLineItem:
    properties:
      amount:
        type: string
      category_id:
        type: string
      id:
        type: string
      note:
        type: string
      tag_ids:
        items:
          type: string
        type: array
    type: object
...
  response.ErrorBody:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
  response.ErrorEnvelope:
    properties:
      error:
        $ref: '#/definitions/response.ErrorBody'
    type: object
host: api.goen.wisteria.local
info:
  contact: {}
  description: REST API for Goen Project
  title: Goen API
  version: "1.0"
paths:
...
  /debts:
    get:
      description: Retrieve a list of debts for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Debt'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List Debts
      tags:
      - Debts
    post:
      consumes:
      - application/json
      description: Create a new debt record linked to a contact
      parameters:
      - description: Debt Creation Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDebtRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Debt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create Debt
      tags:
      - Debts
  /debts/{id}:
    delete:
      description: Delete a debt record by ID
      parameters:
      - description: Debt ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Delete Debt
      tags:
      - Debts
    get:
      description: Retrieve a specific debt by its ID
      parameters:
      - description: Debt ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Debt'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get Debt
      tags:
      - Debts
    patch:
      consumes:
      - application/json
      description: Partially update specific debt properties
      parameters:
      - description: Debt ID
        in: path
        name: id
        required: true
        type: string
      - description: Debt Update Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDebtRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Debt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update Debt
      tags:
      - Debts
  /debts/{id}/payments:
    get:
      description: Retrieve all payments associated with a specific debt
      parameters:
      - description: Debt ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List Debt Payments
      tags:
      - Debts
    post:
      consumes:
      - application/json
      description: Add a new transaction representing a payment toward a specific
        debt
      parameters:
      - description: Debt ID
        in: path
        name: id
        required: true
        type: string
      - description: Debt Payment Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DebtPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Add Debt Payment
      tags:
      - Debts
...
  /group-expenses:
    post:
      consumes:
      - application/json
      description: Create a shared transaction splitting cost among participants
      parameters:
      - description: Group Expense Creation Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGroupExpenseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create Group Expense
      tags:
      - GroupExpenses
  /group-expenses/names:
    get:
      description: Retrieve a list of unique names previously used in group expenses
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List Group Expense Participant Names
      tags:
      - GroupExpenses
  /group-expenses/participants/{transactionId}:
    get:
      description: Retrieve participant portions by an existing transaction ID
      parameters:
      - description: Transaction ID
        in: path
        name: transactionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List Group Expense Participants
      tags:
      - GroupExpenses
  /group-expenses/settle/{participantId}:
    post:
      consumes:
      - application/json
      description: Complete settlement of a specific group expense participant by
        ID
      parameters:
      - description: Participant ID
        in: path
        name: participantId
        required: true
        type: string
      - description: Settlement details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GroupExpenseSettleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Settle Group Expense
      tags:
      - GroupExpenses
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"